requirements.txt
pandas>=2.2.3,<3.0.0
pypdf>=5.1.0,<6.0.0
torch>=2.5.1,<3.0.0
chainlit>=2.0.3,<3.0.0
numpy>=2.2.1,<3.0.0
google-generativeai>=0.8.3,<0.9.0
sentence-transformers>=3.3.1,<4.0.0
llama-index>=0.12.12,<0.13.0
llama-index-llms-gemini>=0.4.4,<0.5.0
llama-index-embeddings-gemini>=0.3.1,<0.4.0
llama-index-embeddings-huggingface>=0.5.1,<0.6.0
pydantic>=2.10.5,<3.0.0
matplotlib>=3.10.0,<4.0.0
llama-index-multi-modal-llms-gemini>=0.5.0,<0.6.0
llama-index-readers-file>=0.4.3,<0.5.0